backend/requirements.txt
alembic==1.18.4
argon2-cffi==25.1.0
openai==1.59.9
email-validator==2.3.0
fastapi==0.136.0
pydantic==2.13.2
pydantic-settings==2.14.0
psycopg==3.3.3
pwdlib==0.3.0
PyJWT==2.12.1
python-multipart==0.0.26
SQLAlchemy==2.0.49
uvicorn==0.44.0
